requirements.txt
requests
notion-client
python-dotenv
pandas
openpyxl
pybit
